SessionGoal now correctly being saved as true once timer completes.

+++ CodingTracker.Common/BusinessInterfaces/CodingSession.cs
@@ -10,10 +10,10 @@ namespace CodingTracker.Common.CodingSessions
         public DateTime? StartTime { get; set; }
         public DateOnly? EndDate { get; set; }
         public DateTime? EndTime { get; set; }
-        public int? DurationSeconds { get; set; }
+        public double? DurationSeconds { get; set; }
         public string DurationHHMM { get; set; } = string.Empty;
         public bool? GoalSet { get; set; }
-        public int? GoalSeconds { get; set; }   
+        public double? GoalSeconds { get; set; }   
         public bool? GoalReached { get; set; }
